refactor(websocket): :art: update websocket folder structure
- moved websocket to service folder - refactored handlers in separated files
