+++ src/App.tsx
@@ -1,7 +1,7 @@
 import { StyledContent, StyledLayout } from '@lib/theme/components/Layout';
 import CoinView from '@/features/coinView/views/CoinView';
 
-import WebSocketComponent from './components/Websocket';
+import WebSocketComponent from './services/Websocket';
 
 function App() {
   return (
