+++ src/services/Websocket.tsx
@@ -1,0 +1,53 @@
+import { useEffect } from 'react';
+
+import { useCoinPairStore } from '@/stores/coinPair';
+
+import { handleL2UpdateMessage, handleSnapshotMessage, handleTickerMessage } from './utils';
+
+const WebSocketComponent = () => {
+  const { selectedPair, pairAsks, pairBids, setPairTicker, setPairAsks, setPairBids } = useCoinPairStore();
+
+  useEffect(() => {
+    const ws = new WebSocket('wss://ws-feed.exchange.coinbase.com');
+    if (selectedPair) {
+      ws.onopen = () => {
+        ws.send(
+          JSON.stringify({
+            type: 'subscribe',
+            product_ids: [selectedPair],
+            channels: ['level2_batch', 'ticker_batch'],
+          }),
+        );
+      };
+
+      ws.onmessage = (event) => {
+        const data = JSON.parse(event.data);
+        if (data.type === 'ticker') {
+          handleTickerMessage({ data, setPairTicker });
+        }
+        if (data.type === 'snapshot') {
+          handleSnapshotMessage({ data, setPairAsks, setPairBids });
+        }
+        if (data.type === 'l2update') {
+          handleL2UpdateMessage({ data, pairAsks, pairBids, setPairAsks, setPairBids });
+        }
+      };
+
+      ws.onerror = (error) => {
+        console.error('WebSocket error:', error);
+      };
+
+      ws.onclose = () => {
+        console.log('WebSocket connection closed');
+      };
+    }
+
+    return () => {
+      ws.close();
+    };
+  }, [selectedPair]);
+
+  return null;
+};
+
+export default WebSocketComponent;
